requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-firebase==1.2
requests==2.7.0
Werkzeug==0.10.4
